requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0
google-cloud-storage==2.10.0
pyjwt==2.8.0
email-validator==2.1.0
bcrypt==4.1.2
alembic==1.13.1
langgraph==0.0.62
langchain==0.1.20
langchain-google-genai==1.0.7
langchain-openai==0.1.8
pinecone-client==3.2.2
openai==1.35.0
numpy==1.24.3
